move helper/logging to internal
remove a dead code file too
