+++ backend/remote-state/azure/sender.go
@@ -6,7 +6,7 @@ import (
 	"net/http/httputil"
 
 	"github.com/Azure/go-autorest/autorest"
-	"github.com/hashicorp/terraform/helper/logging"
+	"github.com/hashicorp/terraform/internal/logging"
 )
 
 func buildSender() autorest.Sender {
