+++ dag/dag_test.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/hashicorp/terraform/tfdiags"
 
-	"github.com/hashicorp/terraform/helper/logging"
+	"github.com/hashicorp/terraform/internal/logging"
 )
 
 func TestMain(m *testing.M) {
