+++ internal/getproviders/http_mirror_source.go
@@ -18,8 +18,8 @@ import (
 	"golang.org/x/net/idna"
 
 	"github.com/hashicorp/terraform/addrs"
-	"github.com/hashicorp/terraform/helper/logging"
 	"github.com/hashicorp/terraform/httpclient"
+	"github.com/hashicorp/terraform/internal/logging"
 	"github.com/hashicorp/terraform/version"
 )
